Added CRUD views for book. Finished api client for backend books

+++ bookstore-backend/bookstore-backend-model/src/main/java/com/training/bookstore/repository/BookRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookRepository extends CrudRepository<BookEntity, Integer> {
+public interface BookRepository extends CrudRepository<BookEntity, Long> {
 
 }
